Refactor offset control into separate `uiBackgroundOffset` module
